requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
django-filter>=23.0,<24.0
psycopg2-binary>=2.9.5,<3.0.0
